Added splitToTest
