+++ tests/Rene.Utils.Core.UnitTest/Extensions/Primitives/DateTimeExtensionsTest.cs
@@ -1,11 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Xunit;
 
-namespace Rene.Utils.Core.UnitTest.Extensions
+namespace Rene.Utils.Core.UnitTest.Extensions.Primitives
 {
     public class DateTimeExtensionsTest
     {
